Create Auto presentation
Pulls data from multiple reports across the network and local drive, updates data where needed, refreshes the pivot data and graphs (saved in a separate reporting spreadsheet) and pushes the results to a powerpoint presentation with MTD and YTD figures against the target KPI's. 

All dates are based off the previous month which can be changed in the 'control' sheet. The network locations have been <redacted> for privacy.

This was my first attempt at auto reports and presentations so the code is very messy, any tips to tidy it up are appreciated!